+++ epa-avert-webapp/src/app/redux/index.ts
@@ -1,20 +1,20 @@
 import { Action, combineReducers } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 // reducers
-import region from 'app/redux/region';
-import panel from 'app/redux/panel';
-import api from 'app/redux/api';
-import rdfs from 'app/redux/rdfs';
-import generation from 'app/redux/generation';
-import so2 from 'app/redux/so2';
-import nox from 'app/redux/nox';
-import co2 from 'app/redux/co2';
-import pm25 from 'app/redux/pm25';
-import eere from 'app/redux/eere';
-import annualDisplacement from 'app/redux/annualDisplacement';
-import stateEmissions from 'app/redux/stateEmissions';
-import monthlyEmissions from 'app/redux/monthlyEmissions';
-import dataDownload from 'app/redux/dataDownload';
+import region from 'app/redux/reducers/region';
+import panel from 'app/redux/reducers/panel';
+import api from 'app/redux/reducers/api';
+import rdfs from 'app/redux/reducers/rdfs';
+import generation from 'app/redux/reducers/generation';
+import so2 from 'app/redux/reducers/so2';
+import nox from 'app/redux/reducers/nox';
+import co2 from 'app/redux/reducers/co2';
+import pm25 from 'app/redux/reducers/pm25';
+import eere from 'app/redux/reducers/eere';
+import annualDisplacement from 'app/redux/reducers/annualDisplacement';
+import stateEmissions from 'app/redux/reducers/stateEmissions';
+import monthlyEmissions from 'app/redux/reducers/monthlyEmissions';
+import dataDownload from 'app/redux/reducers/dataDownload';
 
 const rootReducer = combineReducers({
   region,
@@ -35,50 +35,6 @@ const rootReducer = combineReducers({
 
 export default rootReducer;
 
-// -----------------------------------------------------------------------------
-// types used in other reducers
-// -----------------------------------------------------------------------------
 type RootState = ReturnType<typeof rootReducer>;
 
-type AppThunk = ThunkAction<void, RootState, unknown, Action<string>>;
-
-type DataByMonth = {
-  [MonthNumber in 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12]: number;
-};
-
-type MonthlyChanges = {
-  emissions: {
-    region: DataByMonth;
-    state: {
-      [stateId: string]: DataByMonth;
-    };
-    county: {
-      [stateId: string]: {
-        [countyName: string]: DataByMonth;
-      };
-    };
-  };
-  percentages: {
-    region: DataByMonth;
-    state: {
-      [stateId: string]: DataByMonth;
-    };
-    county: {
-      [stateId: string]: {
-        [countyName: string]: DataByMonth;
-      };
-    };
-  };
-};
-
-type DisplacementData = {
-  original: number;
-  post: number;
-  impact: number;
-  monthlyChanges: MonthlyChanges;
-  stateChanges: {
-    [stateId: string]: number;
-  };
-};
-
-export type { AppThunk, DataByMonth, MonthlyChanges, DisplacementData };
+export type AppThunk = ThunkAction<void, RootState, unknown, Action<string>>;
