+++ epa-avert-webapp/src/app/redux/reducers/generation.ts
@@ -1,43 +1,43 @@
 // reducers
-import { AppThunk, DisplacementData } from 'app/redux/index';
-import { initialPollutantState } from 'app/redux/_common';
+import { AppThunk } from 'app/redux/index';
+import { DisplacementData, initialPollutantState } from 'app/redux/shared';
 // engines
 import { avert } from 'app/engines';
 // action creators
-import { incrementProgress } from 'app/redux/annualDisplacement';
+import { incrementProgress } from 'app/redux/reducers/annualDisplacement';
 
 // action types
-export const REQUEST_CO2 = 'co2/REQUEST_CO2';
-export const RECEIVE_CO2 = 'co2/RECEIVE_CO2';
-export const RECEIVE_ERROR = 'co2/RECEIVE_ERROR';
+export const REQUEST_GENERATION = 'generation/REQUEST_GENERATION';
+export const RECEIVE_GENERATION = 'generation/RECEIVE_GENERATION';
+export const RECEIVE_ERROR = 'generation/RECEIVE_ERROR';
 
-type Co2Action =
+type GenerationAction =
   | {
-      type: typeof REQUEST_CO2;
+      type: typeof REQUEST_GENERATION;
     }
   | {
-      type: typeof RECEIVE_CO2;
+      type: typeof RECEIVE_GENERATION;
       payload: DisplacementData;
     }
   | {
       type: typeof RECEIVE_ERROR;
     };
 
-type Co2State = {
+type GenerationState = {
   isFetching: boolean;
   data: DisplacementData;
   error: boolean;
 };
 
 // reducer
-const initialState: Co2State = initialPollutantState;
+const initialState: GenerationState = initialPollutantState;
 
 export default function reducer(
-  state: Co2State = initialState,
-  action: Co2Action,
-): Co2State {
+  state: GenerationState = initialState,
+  action: GenerationAction,
+): GenerationState {
   switch (action.type) {
-    case REQUEST_CO2:
+    case REQUEST_GENERATION:
       return {
         ...state,
         isFetching: true,
@@ -45,7 +45,7 @@ export default function reducer(
         error: initialState.error,
       };
 
-    case RECEIVE_CO2:
+    case RECEIVE_GENERATION:
       return {
         ...state,
         isFetching: false,
@@ -65,13 +65,13 @@ export default function reducer(
 }
 
 // action creators
-export const fetchCo2 = (): AppThunk => {
+export const fetchGeneration = (): AppThunk => {
   return (dispatch, getState) => {
     const { api } = getState();
 
-    dispatch({ type: REQUEST_CO2 });
+    dispatch({ type: REQUEST_GENERATION });
 
-    // post co2 data for region and receive calculated displacement data
+    // post generation data for region and receive calculated displacement data
     const options = {
       method: 'POST',
       headers: {
@@ -84,12 +84,12 @@ export const fetchCo2 = (): AppThunk => {
       }),
     };
 
-    return fetch(`${api.baseUrl}/api/v1/co2`, options)
+    return fetch(`${api.baseUrl}/api/v1/generation`, options)
       .then((response) => response.json())
       .then((json) => {
         dispatch(incrementProgress());
         dispatch({
-          type: RECEIVE_CO2,
+          type: RECEIVE_GENERATION,
           payload: json,
         });
       })
