+++ epa-avert-webapp/src/app/redux/reducers/so2.ts
@@ -1,43 +1,43 @@
 // reducers
-import { AppThunk, DisplacementData } from 'app/redux/index';
-import { initialPollutantState } from 'app/redux/_common';
+import { AppThunk } from 'app/redux/index';
+import { DisplacementData, initialPollutantState } from 'app/redux/shared';
 // engines
 import { avert } from 'app/engines';
 // action creators
-import { incrementProgress } from 'app/redux/annualDisplacement';
+import { incrementProgress } from 'app/redux/reducers/annualDisplacement';
 
 // action types
-export const REQUEST_NOX = 'nox/REQUEST_NOX';
-export const RECEIVE_NOX = 'nox/RECEIVE_NOX';
-export const RECEIVE_ERROR = 'nox/RECEIVE_ERROR';
+export const REQUEST_SO2 = 'so2/REQUEST_SO2';
+export const RECEIVE_SO2 = 'so2/RECEIVE_SO2';
+export const RECEIVE_ERROR = 'so2/RECEIVE_ERROR';
 
-type NoxAction =
+type So2Action =
   | {
-      type: typeof REQUEST_NOX;
+      type: typeof REQUEST_SO2;
     }
   | {
-      type: typeof RECEIVE_NOX;
+      type: typeof RECEIVE_SO2;
       payload: DisplacementData;
     }
   | {
       type: typeof RECEIVE_ERROR;
     };
 
-type NoxState = {
+type So2State = {
   isFetching: boolean;
   data: DisplacementData;
   error: boolean;
 };
 
 // reducer
-const initialState: NoxState = initialPollutantState;
+const initialState: So2State = initialPollutantState;
 
 export default function reducer(
-  state: NoxState = initialState,
-  action: NoxAction,
-): NoxState {
+  state: So2State = initialState,
+  action: So2Action,
+): So2State {
   switch (action.type) {
-    case REQUEST_NOX:
+    case REQUEST_SO2:
       return {
         ...state,
         isFetching: true,
@@ -45,7 +45,7 @@ export default function reducer(
         error: initialState.error,
       };
 
-    case RECEIVE_NOX:
+    case RECEIVE_SO2:
       return {
         ...state,
         isFetching: false,
@@ -65,13 +65,13 @@ export default function reducer(
 }
 
 // action creators
-export const fetchNox = (): AppThunk => {
+export const fetchSo2 = (): AppThunk => {
   return (dispatch, getState) => {
     const { api } = getState();
 
-    dispatch({ type: REQUEST_NOX });
+    dispatch({ type: REQUEST_SO2 });
 
-    // post nox data for region and receive calculated displacement data
+    // post so2 data for region and receive calculated displacement data
     const options = {
       method: 'POST',
       headers: {
@@ -84,12 +84,12 @@ export const fetchNox = (): AppThunk => {
       }),
     };
 
-    return fetch(`${api.baseUrl}/api/v1/nox`, options)
+    return fetch(`${api.baseUrl}/api/v1/so2`, options)
       .then((response) => response.json())
       .then((json) => {
         dispatch(incrementProgress());
         dispatch({
-          type: RECEIVE_NOX,
+          type: RECEIVE_SO2,
           payload: json,
         });
       })
